# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchCreate": {
        "description": "Queue a deterministic batch run over the seed records.",
        "properties": {
          "refresh": {
            "default": false,
            "title": "Refresh",
            "type": "boolean"
          },
          "size": {
            "default": 50,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Size",
            "type": "integer"
          }
        },
        "title": "BatchCreate",
        "type": "object"
      },
      "BatchSummary": {
        "description": "Aggregate throughput view over one batch parent run.",
        "properties": {
          "average_seconds_per_record": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Seconds Per Record"
          },
          "expected_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Expected Breakdown",
            "type": "object"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "metrics": {
            "additionalProperties": true,
            "title": "Metrics",
            "type": "object"
          },
          "record_count": {
            "title": "Record Count",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "$ref": "#/components/schemas/RunState"
          },
          "states": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "States",
            "type": "object"
          },
          "verdict_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Verdict Breakdown",
            "type": "object"
          },
          "wall_clock_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wall Clock Seconds"
          }
        },
        "required": [
          "run_id",
          "state",
          "record_count"
        ],
        "title": "BatchSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunState": {
        "description": "Lifecycle of an assessment run (architecture run-state machine).",
        "enum": [
          "queued",
          "running",
          "needs_more_data",
          "awaiting_review",
          "complete",
          "failed"
        ],
        "title": "RunState",
        "type": "string"
      },
      "RunSummary": {
        "description": "One run row as it appears in list views and the cockpit table.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "elapsed_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Seconds"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "expected_archetype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Archetype"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "metrics": {
            "additionalProperties": true,
            "title": "Metrics",
            "type": "object"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "parent_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "polygon_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Polygon Id"
          },
          "record_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "$ref": "#/components/schemas/RunState"
          },
          "step_count": {
            "default": 0,
            "title": "Step Count",
            "type": "integer"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "supplier_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Verdict"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "run_id",
          "kind",
          "state",
          "model",
          "created_at"
        ],
        "title": "RunSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Verdict": {
        "description": "Outcome of the deterministic rubric for one supplier/parcel.",
        "enum": [
          "compliant",
          "high_risk",
          "ambiguous"
        ],
        "title": "Verdict",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "TerraSentry API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/batch-runs": {
      "get": {
        "operationId": "list_batch_runs_batch_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BatchSummary"
                  },
                  "title": "Response List Batch Runs Batch Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Batch Runs",
        "tags": [
          "batch"
        ]
      },
      "post": {
        "description": "Queue a batch; records run concurrently on the deterministic assess path.",
        "operationId": "create_batch_batch_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Batch",
        "tags": [
          "batch"
        ]
      }
    },
    "/batch-runs/{batch_run_id}": {
      "get": {
        "operationId": "get_batch_run_batch_runs__batch_run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_run_id",
            "required": true,
            "schema": {
              "title": "Batch Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Batch Run",
        "tags": [
          "batch"
        ]
      }
    },
    "/batch-runs/{batch_run_id}/records": {
      "get": {
        "operationId": "list_batch_records_batch_runs__batch_run_id__records_get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_run_id",
            "required": true,
            "schema": {
              "title": "Batch Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response List Batch Records Batch Runs  Batch Run Id  Records Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Batch Records",
        "tags": [
          "batch"
        ]
      }
    },
    "/batch-runs/{batch_run_id}/stream": {
      "get": {
        "description": "SSE stream of ``progress`` events and a terminal ``done`` event.",
        "operationId": "stream_batch_run_batch_runs__batch_run_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_run_id",
            "required": true,
            "schema": {
              "title": "Batch Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Batch Run",
        "tags": [
          "batch"
        ]
      }
    },
    "/dds/{run_id}": {
      "get": {
        "operationId": "get_dds_dds__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Dds Dds  Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dds",
        "tags": [
          "dds"
        ]
      }
    },
    "/dds/{run_id}/xml": {
      "get": {
        "operationId": "get_dds_xml_dds__run_id__xml_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dds Xml",
        "tags": [
          "dds"
        ]
      }
    },
...
  }
}